requirements.txt
kaggle
python-dotenv
pandas
numpy
openai
streamlit